Reformatted actions
